Tools: ros2: Reduce test duplication with fixture reuse
* Normally these go in conftest and are auto-imported by pytest, but
  LaunchPytest requires any fixtures to be statically visible at import
time and imported directly
