+++ Tools/ros2/ardupilot_dds_tests/test/ardupilot_dds_tests/test_geopose_msg_received.py
@@ -21,15 +21,12 @@ colcon test --packages-select ardupilot_dds_tests \
 
 """
 
-import launch_pytest
 import math
 import pytest
 import rclpy
 import rclpy.node
 from scipy.spatial.transform import Rotation as R
 import threading
-
-from launch import LaunchDescription
 
 from launch_pytest.tools import process as process_tools
 
@@ -38,6 +35,11 @@ from rclpy.qos import QoSReliabilityPolicy
 from rclpy.qos import QoSHistoryPolicy
 
 from geographic_msgs.msg import GeoPoseStamped
+
+from launch_fixtures import (
+    launch_sitl_copter_dds_serial,
+    launch_sitl_copter_dds_udp,
+)
 
 TOPIC = "ap/geopose/filtered"
 # Copied from locations.txt
@@ -63,6 +65,7 @@ def validate_position_cmac(position):
         and math.isclose(position.altitude, CMAC_ABS_ALT, abs_tol=1.0)
     )
 
+
 def wrap_360(angle):
     if angle > 360:
         angle -= 360.0
@@ -71,6 +74,7 @@ def wrap_360(angle):
         angle += 360.0
         return wrap_360(angle)
     return angle
+
 
 def validate_heading_cmac(orientation):
     """
@@ -128,36 +132,6 @@ class GeoPoseListener(rclpy.node.Node):
 
         if validate_heading_cmac(msg.pose.orientation):
             self.orientation_event_object.set()
-
-
-@launch_pytest.fixture
-def launch_sitl_copter_dds_serial(sitl_copter_dds_serial):
-    """Fixture to create the launch description."""
-    sitl_ld, sitl_actions = sitl_copter_dds_serial
-
-    ld = LaunchDescription(
-        [
-            sitl_ld,
-            launch_pytest.actions.ReadyToTest(),
-        ]
-    )
-    actions = sitl_actions
-    yield ld, actions
-
-
-@launch_pytest.fixture
-def launch_sitl_copter_dds_udp(sitl_copter_dds_udp):
-    """Fixture to create the launch description."""
-    sitl_ld, sitl_actions = sitl_copter_dds_udp
-
-    ld = LaunchDescription(
-        [
-            sitl_ld,
-            launch_pytest.actions.ReadyToTest(),
-        ]
-    )
-    actions = sitl_actions
-    yield ld, actions
 
 
 @pytest.mark.launch(fixture=launch_sitl_copter_dds_serial)
